Mudanças para double
Cara, eu mudei as funções para receberem double nos parametros, mas a
função q eu usava para converter de binario para decimal e de decimal
para binarios era da classe Integer, então nao podemos mais usar
agora... Temos que fazer na mão as funções de conversão, eu fiz duas lá,
a getBinario está dando zica e eu nao descobri o que era, a de decimal
ara binario ta funfando perfeitamente, se vc entrar e eu não estiver
online, da uma olhada nisso. Com isso pronto, é so copiar a mesma
estrutura para todas as funções da UCP menos os jumps que vamos ter que
criar ainda o PC

+++ src/Processador/ULA.java
@@ -14,8 +14,8 @@ public class ULA {
 	}
 	
 	
-	public int sub(int rs, int rt){
-		int result = rs - rt;
+	public double sub(double rs, double rt){
+		double result = rs - rt;
 		return result;	
 		
 	}
